dashboard/requirements_dashboard.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=1.5.0
google-cloud-bigquery>=3.13.0
numpy>=1.24.0
